server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
python-jose==3.3.0
pydantic==2.5.2
gunicorn==21.2.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
